+++ GroceryCheckout/products.txt
@@ -1,0 +1,6 @@
+4011 BANANAS 1 0.49 123.2
+4383 MINNEOLAS 1 0.79 187.3
+3144 TANGERINES 1 1.19 135.5
+4028 STRAWBERRIES_PINT 0 0.99 104
+4252 STRAWBERRIES_HALF_CASE 0 3.99 53
+4249 STRAWBERRIES_FULL_CASE 0 7.49 67.
