--- main/serviceInfo.txt
@@ -1,55 +1,0 @@
-1 33.09 96.67 2 ALLEN 69
-4 32.97 96.85 2 CARROLTON 58
-8 32.86 96.7 2 DALLAS 98
-9 33.83 96.7 2 DALLAS 21
-10 32.97 96.74 2 DALLAS 104
-11 32.91 96.71 2 DALLAS 102
-12 32.94 96.78 2 DALLAS 100
-19 33.11 96.8 2 FRISCO 92
-20 33.1 96.84 2 FRISCO 70
-21 33.14 96.85 2 FRISCO 64
-22 33.14 96.89 2 FRISCO 109
-25 32.88 96.6 2 GARLAND 103
-26 32.91 96.64 2 GARLAND 68
-33 33.05 96.62 2 PARKER 97
-35 33.02 96.7 2 PLANO 30
-36 33.04 96.77 2 PLANO 57
-43 32.94 96.73 2 RICHARDSON 60
-44 32.95 96.69 2 RICHARDSON 27
-51 32.9 96.51 2 ROWLETT 20
-52 32.89 96.56 2 ROWLETT 23
-53 32.92 96.57 2 ROWLETT 89
-2 33.11 96.67 1 ALLEN 47
-5 33.04 96.86 1 CARROLTON 100
-13 32.91 96.79 1 DALLAS 44
-14 32.91 96.77 1 DALLAS 106
-15 32.91 96.74 1 DALLAS 74
-16 32.91 96.86 1 DALLAS 93
-17 32.88 96.66 1 DALLAS 35
-23 33.15 96.84 1 FRISCO 77
-27 32.9 96.62 1 GARLAND 85
-28 32.91 96.66 1 GARLAND 107
-29 32.87 96.61 1 GARLAND 91
-31 33.01 96.6 1 MURPHY 40
-37 33.01 96.74 1 PLANO 81
-38 33.01 96.79 1 PLANO 36
-39 33.04 96.71 1 PLANO 29
-40 33.03 96.7 1 PLANO 49
-41 33.04 96.86 1 PLANO 32
-45 32.94 96.73 1 RICHARDSON 88
-46 32.92 96.74 1 RICHARDSON 20
-47 32.93 96.72 1 RICHARDSON 109
-48 32.99 96.75 1 RICHARDSON 84
-49 32.97 96.7 1 RICHARDSON 63
-54 32.91 96.52 1 ROWLETT 63
-0 33.09 96.67 0 ALLEN 46
-3 32.97 96.88 0 CARROLTON 30
-6 32.98 96.78 0 DALLAS 85
-7 32.92 96.81 0 DALLAS 31
-18 33.13 96.82 0 FRISCO 99
-24 32.9 96.66 0 GARLAND 87
-30 33.01 96.6 0 MURPHY 33
-32 33.05 96.62 0 PARKER 58
-34 33.01 96.7 0 PLANO 25
-42 32.94 96.73 0 RICHARDSON 23
-50 32.89 96.57 0 ROWLETT 101
